refid: 3f01d4ef6da9b84171f0072fb74141f02a47adb4

+++ ScreenMeetSDK/Classes/Transport/WebSocket/Model/SMSendTrackPayload.swift
@@ -120,7 +120,7 @@ struct SMSendTrackRtpEncoding: Codable, SocketData {
     var ssrc: Int?
     var active: Bool?
     var dtx: Bool?
-    var maxBitrate: Int?
+    var maxBitrate: Double?
     var networkPriority: Int?
     var rid: String?
